+++ seantis/people/browser/base.py
@@ -2,6 +2,8 @@ from five import grok
 from plone.directives.form import Form
 from z3c.form.interfaces import ActionExecutionError
 from zope.interface import Invalid
+
+from seantis.plonetools import tools
 from seantis.people.interfaces import ISeantisPeopleSpecific
 
 
@@ -9,6 +11,9 @@ class BaseView(grok.View):
 
     grok.baseclass()
     grok.layer(ISeantisPeopleSpecific)
+
+    def translate(self, text):
+        return tools.translator(self.request, 'seantis.people')(text)
 
 
 class BaseForm(Form):