+++ seantis/people/supermodel/schema_columns.py
@@ -1,7 +1,10 @@
 from seantis.plonetools import tools
 from seantis.people.supermodel.security import has_read_access
 from seantis.people.supermodel import (
-    get_selectable_fields, get_columns, get_title_fields
+    get_columns,
+    get_custom_column_titles,
+    get_selectable_fields,
+    get_title_fields,
 )
 
 
@@ -15,31 +18,35 @@ def get_schema_columns(schema, context):
 
 def _get_schema_columns(schema, context, restricted):
     columns = []
+    titles = get_custom_column_titles(schema)
 
-    for fields in get_columns(schema):
+    for ix, fields in enumerate(get_columns(schema)):
         if restricted:
             fields = [
                 f for f in fields if has_read_access(schema, f, context)
             ]
 
         if fields:
-            columns.append(SchemaColumn(schema, fields))
+            columns.append(SchemaColumn(schema, fields, titles[ix]))
 
     return columns
 
 
 class SchemaColumn(object):
 
-    def __init__(self, schema, fields):
+    def __init__(self, schema, fields, custom_title=None):
         self.schema = schema
         self.fields = tools.order_fields_by_schema(fields, schema)
 
-        self.titles = []
-        for field in self.fields:
-            if field in self.schema:
-                self.titles.append(self.schema[field].title)
-            else:
-                self.titles.append(field)
+        if custom_title is not None:
+            self.titles = [custom_title]
+        else:
+            self.titles = []
+            for field in self.fields:
+                if field in self.schema:
+                    self.titles.append(self.schema[field].title)
+                else:
+                    self.titles.append(field)
 
         self.contains_title_field = bool(
             set(self.fields) & set(get_title_fields(self.schema))